PDFBOX-5660: fix cast
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1914334 13f79535-47bb-0310-9956-ffa450edef68

+++ fontbox/src/main/java/org/apache/fontbox/cff/Type1CharStringParser.java
@@ -90,7 +90,7 @@ public class Type1CharStringParser
                             + glyphName + "' of font " + fontName);
                     continue;
                 }
-                int operand = (int) obj;
+                int operand = (int) (Integer) obj;
 
                 if (operand >= 0 && operand < subrs.size())
                 {
